remove the RalixCtrl concept and move to Core methods (injected in global before a controller)
