+++ index.js
@@ -1,7 +1,3 @@
 import RalixApp  from './src/app'
-import RalixCtrl from './src/controller'
 
-export {
-  RalixApp,
-  RalixCtrl
-}
+export { RalixApp }
